Version 1.0: Message the Future
App schedules messages to be sent at a future date & time.
The app sends the message, contact phone number, device phone number and date and time to a DB on an Amazon ec2 instance.
The App has a service that checks the server every 30 seconds to see if the device has any messages to send.